Module type Backend_impl.Backend_impl_common

Parts shared by backend implementations excluding what's already in Backend_any_common, except for Buffer which is duplicated for technical reasons.

include Backend_intf.Buffer

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val c_ptr_to_string : (buffer_ptr -> Ops.prec -> Base.string) Base.option

include sig ... end

type nonrec buffer = buffer_ptr Backend_intf.buffer

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type nonrec ctx_arrays = buffer_ptr Backend_intf.ctx_arrays

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

val is_in_context : Low_level.traced_array -> Base.bool

If true, the node is required to be in the contexts linked with code that uses it.

Should return false for nodes that are virtual, local, or which the backend prefers to access directly from the host.