Module type Backend_intf.Backend_any_common

Parts shared by both assignments-level and lowered-level backend interfaces.

include Buffer

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val c_ptr_to_string : (buffer_ptr -> Ops.prec -> Base.string) Base.option

include sig ... end

type nonrec buffer = buffer_ptr buffer

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type nonrec ctx_arrays = buffer_ptr ctx_arrays

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

val initialize : config -> Base.unit

Initializes a backend before first use. Typically does nothing if the backend is already initialized, but some backends can do some safe cleanups.

val is_initialized : Base.unit -> Base.bool

Returns false if there was no previous initialize call. If it returns false, one must call initialize before using the backend.