Module type Backend_intf.With_buffer_retrieval_and_syncing

type context

type event

val from_host : context -> Tnode.t -> Base.bool

If the tensor node is both hosted and in-context, schedules a copy from host to context and returns true, otherwise returns false.

val to_host : context -> Tnode.t -> Base.bool

If the tensor node is both hosted and in-context, schedules a copy from context to host and returns true, otherwise returns false.

val device_to_device : Tnode.t -> into_merge_buffer:merge_buffer_use -> dst:context -> src:context -> Base.bool

device_to_device tn ~into_merge_buffer ~dst ~src proceeds as follows:

If the node is absent from the src context and either it is present in the dst context or into_merge_buffer is different from No: raises an error.

If the node is absent from dst and into_merge_buffer=No: returns false.

Schedules waiting for writing into the tensor node on src to finish, if any.

If into_merge_buffer=No: schedules a copy of the tensor node from src to dst and updates the writer event for the node.

If into_merge_buffer is different from No: sets on dst the merge buffer source to the given node.

If into_merge_buffer=Streaming, remembers the buffer pointer of the source node to use for streaming.

If into_merge_buffer=Copy, schedules copying from src to the merge buffer of dst's stream, and updates the writer event for the merge buffer.

NOTE: If into_merge_buffer=Streaming, after scheduling the work on dst using the merge buffer but before scheduling work on src that modifies tn, execute will_wait_for src (all_work (get_ctx_stream dst)).