Module Arrayjit.Backend_types

module Lazy = Utils.Lazy

module Debug_runtime = Utils.Debug_runtime

val _get_local_debug_runtime : unit -> (module Minidebug_runtime.Debug_runtime)

type 'buffer_ptr buffer = {

ptr : 'buffer_ptr;

size_in_bytes : Base.int;

}

val sexp_of_buffer : 'buffer_ptr. ('buffer_ptr -> Sexplib0.Sexp.t) -> 'buffer_ptr buffer -> Sexplib0.Sexp.t

module Buffer_types (Buffer_ptr : sig ... end) : sig ... end

module type Buffer = sig ... end

module type Alloc_buffer = sig ... end

module type No_device_buffer_and_copying = sig ... end

module No_device_buffer_and_copying : No_device_buffer_and_copying with type buffer_ptr = Base.unit Ctypes.ptr

type config =

| Only_devices_parallel

| For_parallel_copying

| Most_parallel_streams

For now, we only configure a backend with regard to how many streams it should suggest using (where applicable).

val equal_config : config -> config -> Base.bool

val config_of_sexp : Sexplib0.Sexp.t -> config

val sexp_of_config : config -> Sexplib0.Sexp.t

val only_devices_parallel : config

val for_parallel_copying : config

val most_parallel_streams : config

val is_only_devices_parallel : config -> bool

val is_for_parallel_copying : config -> bool

val is_most_parallel_streams : config -> bool

val only_devices_parallel_val : config -> unit Stdlib.Option.t

val for_parallel_copying_val : config -> unit Stdlib.Option.t

val most_parallel_streams_val : config -> unit Stdlib.Option.t

module Variants_of_config : sig ... end

type merge_buffer_use =

| No

| Streaming

| Copy

val equal_merge_buffer_use : merge_buffer_use -> merge_buffer_use -> Base.bool

val merge_buffer_use_of_sexp : Sexplib0.Sexp.t -> merge_buffer_use

val sexp_of_merge_buffer_use : merge_buffer_use -> Sexplib0.Sexp.t

type param_source =

| Log_file_name

| Merge_buffer

| Param_ptr of Tnode.t

| Static_idx of Indexing.static_symbol

val sexp_of_param_source : param_source -> Sexplib0.Sexp.t

type 'context routine = {

context : 'context;

schedule : Task.t;

bindings : Indexing.lowered_bindings;

name : Base.string;

inputs : Base.Set.M(Arrayjit.Tnode).t;
(*
The materialized read-only and read-before-write (within the routine) non-constant nodes. They are inputs in a broad sense, as they could be recurrent nodes or parameters.
*)

outputs : Base.Set.M(Arrayjit.Tnode).t;
(*
All the materialized nodes written-to by the routine.
*)

}

val sexp_of_routine : 'context. ('context -> Sexplib0.Sexp.t) -> 'context routine -> Sexplib0.Sexp.t

module type Device_config = sig ... end

type ('buffer_ptr, 'device, 'stream_state, 'runner, 'event) stream = {

device : 'device;

state : 'stream_state;

merge_buffer : ('buffer_ptr * Tnode.t) Base.option Base.ref;

unique_name : Base.string;

mutable allocated_buffer : 'buffer_ptr buffer Base.option;

runner : 'runner;

requested_work_for : 'event Base.option Base.Hashtbl.M(Arrayjit.Tnode).t;

}

val sexp_of_stream : 'buffer_ptr 'device 'stream_state 'runner 'event. ('buffer_ptr -> Sexplib0.Sexp.t) -> ('device -> Sexplib0.Sexp.t) -> ('stream_state -> Sexplib0.Sexp.t) -> ('runner -> Sexplib0.Sexp.t) -> ('event -> Sexplib0.Sexp.t) -> ('buffer_ptr, 'device, 'stream_state, 'runner, 'event) stream -> Sexplib0.Sexp.t

module type Device_types = sig ... end

module Stream (Device_config : Device_config) : sig ... end

module type Device = sig ... end

module Device_types (Device_config : Device_config) : sig ... end

module Device (Device_types : Device_types) (Alloc_buffer : Alloc_buffer with type buffer_ptr := Device_types.buffer_ptr and type stream := Device_types.stream) : sig ... end

module type Backend_any_common = sig ... end

Parts shared by both assignments-level and lowered-level backend interfaces.

module type Backend_common = sig ... end

Parts shared by assignments-level backend interfaces.

module type Backend_impl_common = sig ... end

Parts shared by backend implementations excluding what's already in Backend_any_common, except for Buffer which is duplicated for technical reasons.

module type No_device_backend = sig ... end

An intermediate interface for stream-agnostic (typically CPU) backend implementations.

module type Backend_device_common = sig ... end

Parts shared by both assignments-level and lowered-level backend interfaces providing streams and devices.

module type With_buffer_retrieval_and_syncing = sig ... end

module type Backend = sig ... end

module type Lowered_no_device_backend = sig ... end

Lowered-level stream agnostic backend interface: implementation-facing API for CPU backends.

module type No_buffer_retrieval_or_syncing = sig ... end

module type Lowered_backend = sig ... end

Lowered-level backend interface: implementation-facing API for device-based (typically GPU) backends.