Module type Backend_types.Backend_any_common

Parts shared by both assignments-level and lowered-level backend interfaces.

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

type context

val sexp_of_context : context -> Sexplib0.Sexp.t

type stream

type init_info

For backends derived via No_device_backend, this is usually the backend name concatenated with the device or stream number. For Backend, init_info = stream.

val name : Base.string

val initialize : Types.config -> Base.unit

Initializes a backend before first use. Typically does nothing if the backend is already initialized, but some backends can do some safe cleanups.

val is_initialized : Base.unit -> Base.bool

Returns false if there was no previous initialize call. If it returns false, one must call initialize before using the backend.

val init : init_info -> context

val finalize : context -> Base.unit

Finalizes (just) the context.

val alloc_buffer : ?old_buffer:(buffer_ptr * Base.int) -> size_in_bytes:Base.int -> stream -> buffer_ptr