Module type Backend_types.Backend_common

Parts shared by assignments-level backend interfaces.

include Backend_any_common

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

type context

val sexp_of_context : context -> Sexplib0.Sexp.t

type stream

type init_info

For backends derived via No_device_backend, this is usually the backend name concatenated with the device or stream number. For Backend, init_info = stream.

val name : Base.string

val initialize : Types.config -> Base.unit

Initializes a backend before first use. Typically does nothing if the backend is already initialized, but some backends can do some safe cleanups.

val is_initialized : Base.unit -> Base.bool

Returns false if there was no previous initialize call. If it returns false, one must call initialize before using the backend.

val init : init_info -> context

val finalize : context -> Base.unit

Finalizes (just) the context.

val alloc_buffer : ?old_buffer:(buffer_ptr * Base.int) -> size_in_bytes:Base.int -> stream -> buffer_ptr

type routine = context Types.routine

val sexp_of_routine : routine -> Sexplib0.Sexp.t

type code

val sexp_of_code : code -> Sexplib0.Sexp.t

type code_batch

val sexp_of_code_batch : code_batch -> Sexplib0.Sexp.t

val compile : ?shared:Base.bool -> ?name:Base.string -> Indexing.unit_bindings -> Assignments.comp -> code

If ~shared:true (default false), the backend should prefer to do more compile work in a device-and-stream-agnostic way. If ~shared:false, the backend can opt to postpone compiling altogether until link is called, to benefit from more optimizations.

val compile_batch : ?shared:Base.bool -> ?names:Base.string Base.array -> ?occupancy:(name:Base.string -> src_n:Base.int -> Base.bool) -> Indexing.unit_bindings -> Assignments.comp Base.array -> code_batch

Unlike the ~shared parameter, compile_batch vs. compile is mostly about improving the compile time and debugging convenience by generating fewer files -- ideally does not affect execution, but there can be backend-specific differences. Only array entries for which occupancy returns true are included.