Module type Backend_types.Device_types

include Device_config

include Buffer

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val c_ptr_to_string : (buffer_ptr -> Ops.prec -> Base.string) Base.option

include sig ... end

type nonrec buffer = buffer_ptr buffer

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type ctx_arrays = buffer_ptr Base.Map.M(Arrayjit.Tnode).t

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type device

val sexp_of_device : device -> Sexplib0.Sexp.t

type stream_state

val sexp_of_stream_state : stream_state -> Sexplib0.Sexp.t

type runner

val sexp_of_runner : runner -> Sexplib0.Sexp.t

type event

An event tracks if a stream finished computing past a particular point in its schedue. These values are used internally for scheduling across streams of the backend, and can be used for explicit scheduling.

val sexp_of_event : event -> Sexplib0.Sexp.t

type nonrec stream = (buffer_ptr, device, stream_state, runner, event) stream

val sexp_of_stream : stream -> Sexplib0.Sexp.t