Module type Backend_types.Lowered_backend

Lowered-level backend interface: implementation-facing API for device-based (typically GPU) backends.

include No_buffer_retrieval_or_syncing

include Backend_impl_common

type context

include Buffer_ptr

type buffer_ptr

type ctx_arrays = buffer_ptr Base.Map.M(Arrayjit.Tnode).t

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

val ctx_arrays : context -> ctx_arrays

val is_in_context : Low_level.traced_array -> Base.bool

If true, the node is required to be in the contexts linked with code that uses it.

Should return false for nodes that are virtual, local, or which the backend prefers to access directly from the host.

include Backend_device_common with type context := context and type buffer_ptr := buffer_ptr

type device

type event

An event tracks if a stream finished computing past a particular point in its schedue. These values are used internally for scheduling across streams of the backend, and can be used for explicit scheduling.

type stream_state

val sexp_of_stream_state : stream_state -> Sexplib0.Sexp.t

type runner

val sexp_of_runner : runner -> Sexplib0.Sexp.t

type stream = (buffer_ptr, event, device, stream_state, runner) Types.stream

val sexp_of_stream : stream -> Sexplib0.Sexp.t

include Backend_any_common with type buffer_ptr := buffer_ptr and type init_info := stream and type stream := stream with type context := context

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val sexp_of_context : context -> Sexplib0.Sexp.t

val name : Base.string

val initialize : Types.config -> Base.unit

Initializes a backend before first use. Typically does nothing if the backend is already initialized, but some backends can do some safe cleanups.

val is_initialized : Base.unit -> Base.bool

Returns false if there was no previous initialize call. If it returns false, one must call initialize before using the backend.

val init : stream -> context

val finalize : context -> Base.unit

Finalizes (just) the context.

val alloc_buffer : ?old_buffer:(buffer_ptr * Base.int) -> size_in_bytes:Base.int -> stream -> buffer_ptr

val sync : event -> Base.unit

Blocks till the event completes, if it's not done already.

val is_done : event -> Base.bool

Whether the event completed.

val will_wait_for : context -> event -> Base.unit

Schedules waiting for the given event on the context's stream.

NOTE: it should rarely be needed to call will_wait_for explicitly, because it is typically called internally when necessary. But there is one exception, see device_to_device when into_merge_buffer=Streaming.

val get_used_memory : device -> Base.int

Returns (an upper bound of) the memory used for arrays, in bytes.

val await : stream -> Base.unit

Blocks till the stream becomes idle, i.e. synchronizes the stream.

val all_work : stream -> event

Returns the event indicating if any currently running or scheduled computations on the stream have completed.

val is_idle : stream -> Base.bool

Whether the stream is currently waiting for work.

val get_device : ordinal:Base.int -> device

val num_devices : Base.unit -> Base.int

val suggested_num_streams : device -> Base.int

The optimal number of streams for the given device to follow the Types.config strategy passed to No_device_backend.initialize.

val new_stream : device -> stream

val get_ctx_stream : context -> stream

val get_stream_device : stream -> device

val to_ordinal : device -> Base.int

val get_name : stream -> Base.string

val from_host : dst_ptr:buffer_ptr -> dst:context -> Ndarray.t -> Base.unit

Like Backend.from_host, but without synchronization and buffer retrieval.

val to_host : src_ptr:buffer_ptr -> src:context -> Ndarray.t -> Base.unit

Like Backend.to_host, but without synchronization and buffer retrieval.

val device_to_device : Tnode.t -> into_merge_buffer:Types.merge_buffer_use -> dst_ptr:buffer_ptr Base.option -> dst:context -> src_ptr:buffer_ptr -> src:context -> Base.unit

Like Backend.device_to_device, but without synchronization and buffer retrieval. Raises Invalid_argument if into_merge_buffer = No and dst_ptr = None.

type code

val sexp_of_code : code -> Sexplib0.Sexp.t

type code_batch

val sexp_of_code_batch : code_batch -> Sexplib0.Sexp.t

val compile : name:Base.string -> Indexing.unit_bindings -> Low_level.optimized -> code

val compile_batch : names:Base.string Base.option Base.array -> Indexing.unit_bindings -> Low_level.optimized Base.option Base.array -> code_batch

val link : context -> code -> context * Indexing.lowered_bindings * Task.t

val link_batch : context -> code_batch -> context * Indexing.lowered_bindings * Task.t Base.option Base.array

val scheduled_merge_node : stream -> Tnode.t Base.option

scheduled_merge_node stream is the tensor node that would be in the stream's merge buffer right after await stream.