Module type Backend_types.Lowered_backend_common

Parts shared by lowered-level backends excluding what's already in Backend_any_common.

type context

val sexp_of_context : context -> Sexplib0.Sexp.t

type ctx_array

val sexp_of_ctx_array : ctx_array -> Sexplib0.Sexp.t

type ctx_arrays

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type buffer_ptr

val buffer_ptr : ctx_array -> buffer_ptr

val ctx_arrays : context -> ctx_arrays

val get_array : ctx_arrays -> Tnode.t -> ctx_array Base.option

val is_in_context : Low_level.traced_array -> Base.bool

If true, the node is required to be in the contexts linked with code that uses it.

Should return false for nodes that are virtual, local, or which the backend prefers to access directly from the host.