Module Arrayjit.Backends

The collection of the execution backends

val sync_suggested_num_streams : Base.int Base.ref

val reinitialize : (module Backend_types.Backend) -> Backend_types.config -> Base.unit

Initializes the backend, and if it was already initialized, performs garbage collection.

val fresh_backend : ?backend_name:Base.string -> ?config:Backend_types.config -> Base.unit -> (module Backend_types.Backend)

Reinitializes and returns a backend corresponding to backend_name, or if omitted, selected via the global backend setting. See reinitialize.