Module Backends.Cuda_backend

include No_device_backend with type context = Cuda_backend.context

type code

val sexp_of_code : code -> Sexplib0.Sexp.t

type code_batch

val sexp_of_code_batch : code_batch -> Sexplib0.Sexp.t

type context = Cuda_backend.context

val sexp_of_context : context -> Sexplib0.Sexp.t

type nonrec routine = context routine

val sexp_of_routine : routine -> Sexplib0.Sexp.t

val name : Base.string

val initialize : config -> Base.unit

val is_initialized : Base.unit -> Base.bool

val finalize : context -> Base.unit

Finalizes (just) the context.

val compile : ?shared:Base.bool -> ?name:Base.string -> Indexing.unit_bindings -> Assignments.t -> code

If ~shared:true (default false), the backend should prefer to do more compile work in a device-agnostic way. If ~shared:false, the backend can opt to postpone compiling altogether until link is called, to benefit from more optimizations.

val compile_batch : ?shared:Base.bool -> ?names:Base.string Base.array -> ?occupancy:(name:Base.string -> src_n:Base.int -> Base.bool) -> Indexing.unit_bindings -> Assignments.t Base.array -> code_batch

Unlike the ~shared parameter, compile_batch vs. compile is mostly about improving the compile time and debugging convenience by generating fewer files -- ideally does not affect execution, but there can be backend-specific differences. Only array entries for which occupancy returns true are included.

val link : context -> code -> routine

Returns the routine for the code's procedure, in a new context derived from the given context.

val link_batch : context -> code_batch -> routine Base.option Base.array

Returns the routines for the procedures included in the code batch. All returned routines share the same new context.

val unsafe_cleanup : ?unsafe_shutdown:Base.bool -> Base.unit -> Base.unit

Cleans up all work on a backend. If ~unsafe_shutdown:true, releases resources, potentially making the backend unusable.

val from_host : ?rt:(module Minidebug_runtime.Debug_runtime) -> context -> Tnode.t -> Base.unit

If the array is both hosted and in-context, schedules a copy from host to context, otherwise does nothing. NOTE: when run for a device, it's the caller's responsibility to synchronize the device before the host's data is overwritten.

val to_host : ?rt:(module Minidebug_runtime.Debug_runtime) -> context -> Tnode.t -> Base.unit

If the array is both hosted and in-context, schedules a copy from context to host, otherwise does nothing. NOTE: when run for a device, it's the caller's responsibility to synchronize the device before the host's data is read.

val device_to_device : ?rt:(module Minidebug_runtime.Debug_runtime) -> Tnode.t -> into_merge_buffer:Base.bool -> dst:context -> src:context -> Base.unit

If the array is present in the src context, and in the dst context in case ~into_merge_buffer:false, schedules a copy of the tensor node from the device of src to the device of dst; otherwise, if ~into_merge_buffer:true, unsets the merge buffer source on dst. If ~into_merge_buffer:false, copies into the given node on dst; if ~into_merge_buffer:true, sets on dst the merge buffer source to the given node and handles the transfer using it. NOTE: when run for a device, it's the caller's responsibility to synchronize the src device, if needed, before calling device_to_device, and the dst device afterward, according to physical_merge_buffers, before any computations on the src device overwrite the node.

val physical_merge_buffers : Base.bool

If physical_merge_buffers = true, the src node data is not needed after the task scheduled by device_to_device finishes. Otherwise, the src node data may be needed till after the tasks computing with the merge buffer finish.

type physical_device

type device

val init : device -> context

val await : device -> Base.unit

Blocks till the device becomes idle, i.e. synchronizes the device.

val is_idle : device -> Base.bool

Whether the device is currently waiting for work.

val sexp_of_device : device -> Base.Sexp.t

val get_device : ordinal:Base.int -> physical_device

val num_physical_devices : Base.unit -> Base.int

val suggested_num_virtual_devices : physical_device -> Base.int

The optimal number of virtual devices for the given physical device to follow the No_device_backend.config strategy passed to No_device_backend.initialize.

val new_virtual_device : physical_device -> device

val get_ctx_device : context -> device

val get_physical_device : device -> physical_device

val to_ordinal : physical_device -> Base.int

val to_subordinal : device -> Base.int

val get_name : device -> Base.string