Parameter Lowered_backend.Device

type stream

val sexp_of_stream : stream -> Sexplib0.Sexp.t

include Backend_types.Lowered_backend_common with type config := Backend_types.Types.config and type stream := stream and type init_info := stream

type context

type ctx_array

val sexp_of_ctx_array : ctx_array -> Sexplib0.Sexp.t

type ctx_arrays

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val buffer_ptr : ctx_array -> buffer_ptr

val alloc_buffer : ?old_buffer:(buffer_ptr * Base.int) -> size_in_bytes:Base.int -> stream -> buffer_ptr

val ctx_arrays : context -> ctx_arrays

val get_array : ctx_arrays -> Tnode.t -> ctx_array Base.option

val is_in_context : Low_level.traced_array -> Base.bool

If true, the node is required to be in the contexts linked with code that uses it.

Should return false for nodes that are virtual, local, or which the backend prefers to access directly from the host.

val initialize : Backend_types.Types.config -> Base.unit

val is_initialized : Base.unit -> Base.bool

val init : stream -> context

val finalize : context -> Base.unit

val unsafe_cleanup : Base.unit -> Base.unit

val name : Base.string

type code

val sexp_of_code : code -> Sexplib0.Sexp.t

type code_batch

val sexp_of_code_batch : code_batch -> Sexplib0.Sexp.t

type event

val sync : event -> Base.unit

val is_done : event -> Base.bool

val work_for : context -> Tnode.t -> event Base.option

val will_wait_for : context -> event -> Base.unit

val sexp_of_context : context -> Sexplib.Sexp.t

val compile : name:Base.string -> Indexing.unit_bindings -> Low_level.optimized -> code

val compile_batch : names:Base.string Base.option Base.array -> Indexing.unit_bindings -> Low_level.optimized Base.option Base.array -> code_batch

val link : context -> code -> context * Indexing.lowered_bindings * Task.t

val link_batch : context -> code_batch -> context * Indexing.lowered_bindings * Task.t Base.option Base.array

val from_host : context -> Tnode.t -> Base.bool

If the array is both hosted and in-context, copies from host to context.

val to_host : context -> Tnode.t -> Base.bool

If the array is both hosted and in-context, copies from context to host.

val device_to_device : Tnode.t -> into_merge_buffer:Backend_types.Types.merge_buffer_use -> dst:context -> src:context -> Base.bool

See Backend.device_to_device.

type device

val get_used_memory : device -> Base.int

Returns (an upper bound of) the memory used for arrays, in bytes.

val await : stream -> Base.unit

val is_idle : stream -> Base.bool

val all_work : stream -> event

val scheduled_merge_node : stream -> Tnode.t Base.option

scheduled_merge_node stream is the tensor node that would be in the stream's merge buffer right after await stream.

val num_devices : Base.unit -> Base.int

val suggested_num_streams : device -> Base.int

val get_device : ordinal:Base.int -> device

val get_stream_device : stream -> device

val new_stream : device -> stream

val get_ctx_stream : context -> stream

val get_name : stream -> Base.string

val to_ordinal : device -> Base.int

val to_subordinal : stream -> Base.int