Parameter Lowered_no_device_backend.Backend

include Backend_types.Lowered_backend_common with type stream := Base.unit and type config := Base.unit and type init_info := Base.string

type context

val sexp_of_context : context -> Sexplib0.Sexp.t

type ctx_array

val sexp_of_ctx_array : ctx_array -> Sexplib0.Sexp.t

type ctx_arrays

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val buffer_ptr : ctx_array -> buffer_ptr

val alloc_buffer : ?old_buffer:(buffer_ptr * Base.int) -> size_in_bytes:Base.int -> Base.unit -> buffer_ptr

val ctx_arrays : context -> ctx_arrays

val get_array : ctx_arrays -> Tnode.t -> ctx_array Base.option

val is_in_context : Low_level.traced_array -> Base.bool

If true, the node is required to be in the contexts linked with code that uses it.

Should return false for nodes that are virtual, local, or which the backend prefers to access directly from the host.

val initialize : Base.unit -> Base.unit

val is_initialized : Base.unit -> Base.bool

val init : Base.string -> context

val finalize : context -> Base.unit

val unsafe_cleanup : Base.unit -> Base.unit

val name : Base.string

type procedure

val sexp_of_procedure : procedure -> Sexplib0.Sexp.t

val compile : name:Base.string -> opt_ctx_arrays:ctx_arrays Base.option -> Indexing.unit_bindings -> Low_level.optimized -> procedure

val compile_batch : names:Base.string Base.option Base.array -> opt_ctx_arrays:ctx_arrays Base.option -> Indexing.unit_bindings -> Low_level.optimized Base.option Base.array -> ctx_arrays Base.option * procedure Base.option Base.array

val link_compiled : merge_buffer:(buffer_ptr * Tnode.t) Base.option Base.ref -> context -> procedure -> context * Indexing.lowered_bindings * Task.t * Base.string

val to_buffer : Tnode.t -> dst:buffer_ptr -> src:context -> Base.unit

val host_to_buffer : Ndarray.t -> dst:buffer_ptr -> Base.unit

val buffer_to_host : Ndarray.t -> src:buffer_ptr -> Base.unit