Parameter Multicore_backend.Backend

type code

val sexp_of_code : code -> Sexplib0.Sexp.t

type context

val sexp_of_context : context -> Sexplib0.Sexp.t

type nonrec routine = context routine

val sexp_of_routine : routine -> Sexplib0.Sexp.t

val name : Base.string

val initialize : Base.unit -> Base.unit

val is_initialized : Base.unit -> Base.bool

val init : label:Base.string -> context

label is usually the backend name concatenated with the device number.

val finalize : context -> Base.unit

Finalizes (just) the context.

val compile : ?shared:Base.bool -> ?name:Base.string -> Indexing.unit_bindings -> Assignments.t -> code

If ~shared:true (default false), the backend should prefer to do more compile work in a device-agnostic way. If ~shared:false, the backend can opt to postpone compiling altogether until link is called, to benefit from more optimizations.

val compile_batch : ?shared:Base.bool -> ?names:Base.string Base.array -> Indexing.unit_bindings -> Assignments.t Base.array -> code Base.array

Unlike the ~shared parameter, compile_batch vs. compile is mostly about improving the compile time and debugging convenience by generating fewer files -- ideally does not affect execution, but there can be backend-specific differences.

val link : context -> code -> routine

val unsafe_cleanup : ?unsafe_shutdown:Base.bool -> Base.unit -> Base.unit

Cleans up all work on a backend. If ~unsafe_shutdown:true, releases resources, potentially making the backend unusable.

val from_host : context -> Tnode.t -> Base.bool

If the array is both hosted and in-context, copies from host to context and returns true.

val to_host : context -> Tnode.t -> Base.bool

If the array is both hosted and in-context, copies from context to host and returns true.

val merge : ?name_prefix:Base.string -> Tnode.t -> accum:Ops.binop -> src:context -> code Base.option

Merges the array from the source context into the destination context: dst =: dst accum src. If the array is hosted, its state on host is undefined after this operation. (A backend may choose to use the host array as a buffer, if that is beneficial.) name_prefix is prepended to the jitted function's name. Returns None if the array is not in the context.

val merge_batch : ?name_prefixes:Base.string Base.array -> occupancy:(Tnode.t -> src_n:Base.int -> src:context -> Utils.requirement) -> Tnode.t Base.list -> accum:Ops.binop -> srcs:context Base.array -> (Tnode.t, code Base.option Base.array) Base.Hashtbl.t

merge_batch vs. merge is purely about improving the compile time (does not affect execution).