Parameter C_syntax.B

type buffer_ptr

val procs : (Low_level.optimized * buffer_ptr Backend_intf.ctx_arrays Base.option) Base.array

The low-level prcedure to compile, and the arrays of the context it will be linked to if not shared and already known.

val hardcoded_context_ptr : (buffer_ptr -> Ops.prec -> Base.string) Base.option

val unified_memory : Base.bool

val host_ptrs_for_readonly : Base.bool

val logs_to_stdout : Base.bool

val main_kernel_prefix : Base.string

val kernel_prep_line : Base.string

val include_lines : Base.string Base.list

val typ_of_prec : Ops.prec -> Base.string

val binop_syntax : Ops.prec -> Ops.binop -> Base.string * Base.string * Base.string

val unop_syntax : Ops.prec -> Ops.unop -> Base.string * Base.string

val convert_precision : from:Ops.prec -> to_:Ops.prec -> Base.string * Base.string