Module Arrayjit.Cc_backend

module Lazy = Utils.Lazy

module Debug_runtime = Utils.Debug_runtime

val name : string

val optimization_level : unit -> Base.Int.t

val compiler_command : unit -> Base.String.t

Currently unused, backend behaves as if config is always `Physical_devices_only.

type mem_properties =

| Local_only
(*
The array is only needed for a local computation, is allocated on the stack.
*)

| From_context
(*
The array has a copy allocated per-cpu-device, may or may not exist on the host.
*)

| Constant_from_host
(*
The array is read directly from the host.
*)

val mem_properties_of_sexp : Sexplib0.Sexp.t -> mem_properties

val sexp_of_mem_properties : mem_properties -> Sexplib0.Sexp.t

val equal_mem_properties : mem_properties -> mem_properties -> Base.bool

val compare_mem_properties : mem_properties -> mem_properties -> Base.int

val local_only : mem_properties

val from_context : mem_properties

val constant_from_host : mem_properties

val is_local_only : mem_properties -> bool

val is_from_context : mem_properties -> bool

val is_constant_from_host : mem_properties -> bool

val local_only_val : mem_properties -> unit Stdlib.Option.t

val from_context_val : mem_properties -> unit Stdlib.Option.t

val constant_from_host_val : mem_properties -> unit Stdlib.Option.t

module Variants_of_mem_properties : sig ... end

module Tn = Tnode

type ctx_array = Ndarray.t

val sexp_of_ctx_array : ctx_array -> Sexplib0.Sexp.t

type ctx_arrays = ctx_array Base.Map.M(Tn).t

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type context = {

label : Base.string;

arrays : ctx_arrays;

}

val sexp_of_context : context -> Sexplib0.Sexp.t

val ctx_arrays : context -> ctx_arrays

type buffer_ptr = ctx_array

Alternative approach:

type buffer_ptr = unit Ctypes_static.ptr

let sexp_of_buffer_ptr ptr = Sexp.Atom (Ops.ptr_to_string ptr Ops.Void_prec)
let buffer_ptr ctx_array = Ndarray.get_voidptr ctx_array

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val buffer_ptr : 'a -> 'b

val alloc_buffer : ?old_buffer:('a * int) -> size_in_bytes:int -> unit -> 'b

val to_buffer : ?rt:'a -> Tn.t -> dst:Ndarray.t -> src:context -> unit

val host_to_buffer : ?rt:'a -> Ndarray.t -> dst:Ndarray.t -> unit

val buffer_to_host : ?rt:'a -> Ndarray.t -> src:Ndarray.t -> unit

val unsafe_cleanup : ?unsafe_shutdown:bool -> unit -> Base.unit

val is_initialized : unit -> bool

val initialize : unit -> Base.unit

val finalize : 'a -> unit

val init : label:Base.string -> context

type procedure = {

lowered : Low_level.optimized;

bindings : Indexing.unit_bindings;

name : Base.string;

result : Dl.library;

params : (Base.string * Backend_utils.Types.param_source) Base.list;

opt_ctx_arrays : Ndarray.t Base.Map.M(Tn).t Base.option;

}

val sexp_of_procedure : procedure -> Sexplib0.Sexp.t

val expected_merge_node : procedure -> Arrayjit.Low_level.Tn.t Base.option

val is_builtin_op : Ops.binop -> bool

val pp_comma : Stdlib.Format.formatter -> unit -> unit

val pp_index : Stdlib.Format.formatter -> Indexing.symbol -> unit

val pp_index_axis : Stdlib.Format.formatter -> Indexing.axis_index -> unit

val pp_array_offset : Stdlib.Format.formatter -> (Indexing.axis_index Base.Array.t * int Base.Array.t) -> unit

val array_offset_to_string : (Indexing.axis_index Base.Array.t * int Base.Array.t) -> string

val is_in_context : Low_level.traced_array -> bool

val header_sep : Re.re

val compile : name:Base.string -> opt_ctx_arrays: (Arrayjit.Low_level.Tn.t Base.Hashtbl.key, Ndarray.t, Tnode.comparator_witness) Base.Map.t Base.Option.t -> (Base.unit -> Base.unit) Indexing.bindings -> Low_level.optimized -> procedure

val compile_batch : names:Base.String.t Base.Option.t Base.Array.t -> opt_ctx_arrays: (Arrayjit.Low_level.Tn.t Base.Hashtbl.key, Ndarray.t, Tnode.comparator_witness) Base.Map.t Base.Option.t -> (Base.unit -> Base.unit) Indexing.bindings -> Low_level.optimized Base.option Base.array -> (Arrayjit.Low_level.Tn.t Base.Hashtbl.key, Ndarray.t, Tnode.comparator_witness) Base.Map.t Base.Option.t * procedure Base.Option.t Base.Array.t

val link_compiled : merge_buffer:Ndarray.t Base.option Base.ref -> context -> procedure -> context * (Indexing.static_symbol * Base.int Base.ref) Base.List.t * Tn.task * Base.string