Module Arrayjit.Gccjit_backend

module Lazy = Utils.Lazy

module Debug_runtime = Utils.Debug_runtime

val name : string

val optimization_level : int Base.ref

type config = [

| `Physical_devices_only

| `For_parallel_copying

| `Most_parallel_devices

]

Currently unused, backend behaves as if config is always `Physical_devices_only.

val equal_config : config -> config -> Base.bool

val __config_of_sexp__ : Sexplib0.Sexp.t -> config

val config_of_sexp : Sexplib0.Sexp.t -> config

val sexp_of_config : config -> Sexplib0.Sexp.t

val physical_devices_only : [> `Physical_devices_only ]

val for_parallel_copying : [> `For_parallel_copying ]

val most_parallel_devices : [> `Most_parallel_devices ]

val is_physical_devices_only : [> `Physical_devices_only ] -> bool

val is_for_parallel_copying : [> `For_parallel_copying ] -> bool

val is_most_parallel_devices : [> `Most_parallel_devices ] -> bool

val physical_devices_only_val : [> `Physical_devices_only ] -> unit Stdlib.Option.t

val for_parallel_copying_val : [> `For_parallel_copying ] -> unit Stdlib.Option.t

val most_parallel_devices_val : [> `Most_parallel_devices ] -> unit Stdlib.Option.t

module Variants_of_config : sig ... end

type mem_properties =

| Local_only
(*
The array is only needed for a local computation, is allocated on the stack.
*)

| From_context
(*
The array has a copy allocated per-cpu-device, may or may not exist on the host.
*)

| Constant_from_host
(*
The array is read directly from the host.
*)

val mem_properties_of_sexp : Sexplib0.Sexp.t -> mem_properties

val sexp_of_mem_properties : mem_properties -> Sexplib0.Sexp.t

val equal_mem_properties : mem_properties -> mem_properties -> Base.bool

val compare_mem_properties : mem_properties -> mem_properties -> Base.int

val local_only : mem_properties

val from_context : mem_properties

val constant_from_host : mem_properties

val is_local_only : mem_properties -> bool

val is_from_context : mem_properties -> bool

val is_constant_from_host : mem_properties -> bool

val local_only_val : mem_properties -> unit Stdlib.Option.t

val from_context_val : mem_properties -> unit Stdlib.Option.t

val constant_from_host_val : mem_properties -> unit Stdlib.Option.t

module Variants_of_mem_properties : sig ... end

val root_ctx : Gccjit.context option Base.ref

module Tn = Tnode

type ctx_arrays = Ndarray.t Base.Map.M(Tn).t

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type context = {

label : Base.string;

arrays : ctx_arrays;

result : Gccjit.result Base.option;

}

val sexp_of_context : context -> Sexplib0.Sexp.t

val ctx_arrays : context -> ctx_arrays

val unsafe_cleanup : ?unsafe_shutdown:bool -> unit -> Base.unit

val is_initialized : unit -> bool

val initialize : unit -> Base.unit

val finalize : context -> unit

val init : label:Base.string -> context

type tn_info = {

tn : Tn.t;
(*
The original array.
*)

mutable ptr : Gccjit.rvalue Lazy.t;
(*
Pointer to the first value of the associated array.

if mem = Constant_from_host, the pointer to the first element of the hosted Ndarray,

if mem = From_context, either a pointer to Ndarray from context.arrays when ~shared:false, or the function parameter when ~shared:true,

if mem = Local_only, the address of the on-the-stack array.

*)

mem : mem_properties;

dims : Base.int Base.array;

size_in_bytes : Base.int;

num_typ : Gccjit.type_;
(*
The type of the stored values: short (precision Half), float (precision Single), double (precision Double).
*)

prec : Ops.prec;

zero_initialized : Base.bool;

}

val sexp_of_tn_info : tn_info -> Sexplib0.Sexp.t

type info = {

ctx : Gccjit.context;

func : Gccjit.function_;

traced_store : Low_level.traced_store;

init_block : Gccjit.block;

nodes : (Tn.t, tn_info) Base.Hashtbl.t;

}

val sexp_of_info : info -> Sexplib0.Sexp.t

type param_source =

| Log_file_name

| Param_ptr of Tn.t

| Static_idx of Indexing.static_symbol

val sexp_of_param_source : param_source -> Sexplib0.Sexp.t

type procedure = {

info : info;

bindings : Indexing.unit_bindings;

name : Base.string;

result : Gccjit.result;

opt_ctx_arrays : Ndarray.t Base.Map.M(Tn).t Base.option;

params : param_source Base.list;

}

val sexp_of_procedure : procedure -> Sexplib0.Sexp.t

type gccjit_param = Gccjit.param

type gccjit_lvalue = Gccjit.lvalue

type gccjit_rvalue = Gccjit.rvalue

val sexp_of_gccjit_param : Gccjit.param -> Base.Sexp.t

val sexp_of_gccjit_lvalue : Gccjit.lvalue -> Base.Sexp.t

val sexp_of_gccjit_rvalue : Gccjit.rvalue -> Base.Sexp.t

type ctx_nodes =

| Ctx_arrays of Ndarray.t Base.Map.M(Tn).t Base.ref

| Param_ptrs of (gccjit_param * param_source) Base.list Base.ref

val sexp_of_ctx_nodes : ctx_nodes -> Sexplib0.Sexp.t

val jit_array_offset : Gccjit.context -> idcs:Gccjit.rvalue Base.Array.t -> dims:int Base.Array.t -> Gccjit.rvalue

val zero_out : Gccjit.context -> Gccjit.block -> tn_info -> unit

val get_c_ptr : Gccjit.context -> Gccjit.type_ -> ('a, 'b, 'c) Bigarray_compat.Genarray.t -> Gccjit.rvalue

val prepare_node : debug_log_zero_out:(Gccjit.block -> tn_info -> unit) -> Gccjit.context -> (Tn.t, tn_info) Base.Hashtbl.t -> (Arrayjit.Low_level.Tn.t, Low_level.traced_array) Base.Hashtbl.t -> ctx_nodes -> (Gccjit.block -> Gccjit.function_ -> Base.unit) list Base.ref -> Tn.t -> unit

val prec_to_kind : Ops.prec -> Gccjit.Type.type_kind

val is_builtin_op : Ops.binop -> bool

val builtin_op : Ops.binop -> Gccjit.binary_op

val node_debug_name : tn_info -> Base.String.t

val debug_log_zero_out : Gccjit.context -> (Gccjit.rvalue * Gccjit.function_ * Gccjit.function_) option Lazy.t -> Gccjit.block -> tn_info -> unit

val debug_log_index : Gccjit.context -> (Gccjit.rvalue * Gccjit.function_ * Gccjit.function_) option -> Gccjit.block -> string -> Gccjit.rvalue -> unit

val compile_main : name:Base.String.t -> log_functions:(Gccjit.rvalue * Gccjit.function_ * Gccjit.function_) option -> env:(Indexing.symbol, Gccjit.rvalue, 'a) Base.Map.t -> info -> Gccjit.function_ -> Gccjit.block -> Low_level.t -> Gccjit.block

val prepare_nodes : Gccjit.context -> log_functions: (Gccjit.rvalue * Gccjit.function_ * Gccjit.function_) option Lazy.t -> (Tn.t, tn_info) Base.Hashtbl.t -> (Arrayjit.Low_level.Tn.t, Low_level.traced_array) Base.Hashtbl.t -> ctx_nodes -> (Gccjit.block -> Gccjit.function_ -> Base.unit) list Base.ref -> Low_level.t -> unit

val compile_proc : name:Base.String.t -> opt_ctx_arrays: (Tn.t, Ndarray.t, Tn.comparator_witness) Base.Map.t Base.Option.t -> Gccjit.context -> 'a Indexing.bindings -> Low_level.optimized -> info * Ndarray.t Base.Map.M(Tn).t * (gccjit_param * param_source) Base.list

val header_sep : Re.re

val compile : name:Base.string -> opt_ctx_arrays: (Tn.t, Ndarray.t, Tn.comparator_witness) Base.Map.t Base.Option.t -> (Base.unit -> Base.unit) Indexing.bindings -> Low_level.optimized -> procedure

val compile_batch : names:Base.string Base.option Base.array -> opt_ctx_arrays: (Tn.t, Ndarray.t, Tn.comparator_witness) Base.Map.t Base.Option.t -> (Base.unit -> Base.unit) Indexing.bindings -> Low_level.optimized Base.option Base.array -> procedure Base.Option.t Base.Array.t

val link_compiled : context -> procedure -> context * (Indexing.static_symbol * Base.int Base.ref) Base.List.t * Tn.task * Base.string

val from_host : ?rt:(module Minidebug_runtime.Debug_runtime) -> context -> Tn.t -> Base.unit

val to_host : ?rt:(module Minidebug_runtime.Debug_runtime) -> context -> Tn.t -> Base.unit

val device_to_device : ?rt:(module Minidebug_runtime.Debug_runtime) -> Tn.t -> into_merge_buffer:bool -> dst:context -> src:context -> unit

val physical_merge_buffers : bool