Module Arrayjit.Low_level

A for-loop-based array language and backend-agnostic optimization

module Scope_id : sig ... end

type scope_id = Scope_id.t = {

tn : Tnode.t;

scope_id : Base.int;

}

val sexp_of_scope_id : scope_id -> Sexplib0.Sexp.t

val equal_scope_id : scope_id -> scope_id -> Base.bool

val hash_fold_scope_id : Ppx_hash_lib.Std.Hash.state -> scope_id -> Ppx_hash_lib.Std.Hash.state

val hash_scope_id : scope_id -> Ppx_hash_lib.Std.Hash.hash_value

val compare_scope_id : scope_id -> scope_id -> Base.int

Low-level representation

type t =

| Noop

| Comment of Base.string

| Staged_compilation of Base.unit -> Base.unit

| Seq of t * t

| For_loop of {

index : Indexing.symbol;

from_ : Base.int;

to_ : Base.int;

body : t;

trace_it : Base.bool;

}

| Zero_out of Tnode.t

| Set of {

tn : Tnode.t;

idcs : Indexing.axis_index Base.array;

llv : float_t;

mutable debug : Base.string;

}

| Set_local of scope_id * float_t

Cases: t -- code, float_t -- single number at some precision.

and float_t =

| Local_scope of {

id : scope_id;

body : t;

orig_indices : Indexing.axis_index Base.array;

}

| Get_local of scope_id

| Get_global of Ops.global_identifier * Indexing.axis_index Base.array Base.option

| Get of Tnode.t * Indexing.axis_index Base.array

| Binop of Ops.binop * float_t * float_t

| Unop of Ops.unop * float_t

| Constant of Base.float

| Embed_index of Indexing.axis_index

val sexp_of_t : t -> Sexplib0.Sexp.t

val sexp_of_float_t : float_t -> Sexplib0.Sexp.t

val equal : t -> t -> Base.bool

val equal_float_t : float_t -> float_t -> Base.bool

val compare : t -> t -> Base.int

val compare_float_t : float_t -> float_t -> Base.int

val binop : op:Ops.binop -> rhs1:float_t -> rhs2:float_t -> float_t

val unop : op:Ops.unop -> rhs:float_t -> float_t

val flat_lines : t Base.list -> t Base.list

val unflat_lines : t Base.list -> t

val loop_over_dims : Base.int Base.array -> body:(Indexing.axis_index Base.array -> t) -> t

Optimization

type virtualize_settings = {

mutable enable_device_only : Base.bool;

mutable max_visits : Base.int;

mutable max_tracing_dim : Base.int;

mutable inline_scalar_constexprs : Base.bool;

}

val virtualize_settings : virtualize_settings

type visits =

| Visits of Base.int

| Recurrent
(*
A Recurrent visit is when there is an access prior to any assignment in an update.
*)

val sexp_of_visits : visits -> Sexplib0.Sexp.t

val visits_of_sexp : Sexplib0.Sexp.t -> visits

val equal_visits : visits -> visits -> Base.bool

val visits : Base.int -> visits

val recurrent : visits

val is_visits : visits -> Base.bool

val is_recurrent : visits -> Base.bool

val visits_val : visits -> Base.int Base.option

val recurrent_val : visits -> Base.unit Base.option

module Variants_of_visits : sig ... end

type traced_array = {

tn : Tnode.t;

mutable computations : (Indexing.axis_index Base.array Base.option * t) Base.list;
(*
The computations (of the tensor node) are retrieved for optimization just as they are populated, so that the inlined code corresponds precisely to the changes to the arrays that would happen up till that point. Within the code blocks paired with an index tuple, all assignments and accesses must happen via the index tuple; if this is not the case for some assignment, the node cannot be virtual. Currently, we only allow for-loop symbols in assignment indices of virtual nodes.
*)

assignments : Base.int Base.array Base.Hash_set.t;

accesses : (Base.int Base.array, visits) Base.Hashtbl.t;

mutable zero_initialized : Base.bool;

mutable zeroed_out : Base.bool;

mutable read_before_write : Base.bool;
(*
The node is read before it is written (i.e. it is recurrent).
*)

mutable read_only : Base.bool;

mutable is_scalar_constexpr : Base.bool;
(*
True only if the tensor node has all axes of dimension 1, is either zeroed-out or assigned before accessed, is assigned at most once, and from an expression involving only constants or tensor nodes that were at the time is_scalar_constexpr.
*)

}

val sexp_of_traced_array : traced_array -> Sexplib0.Sexp.t

val get_node : (Tnode.t, traced_array) Base.Hashtbl.t -> Tnode.t -> traced_array

val optimize_integer_pow : Base.bool Base.ref

type traced_store = (Tnode.t, traced_array) Base.Hashtbl.t

val sexp_of_traced_store : traced_store -> Sexplib0.Sexp.t

type optimized = {

traced_store : traced_store;

llc : t;

merge_node : Tnode.t Base.option;

}

val sexp_of_optimized : optimized -> Sexplib0.Sexp.t

val optimize : unoptim_ll_source:Stdlib.Format.formatter Base.option -> ll_source:Stdlib.Format.formatter Base.option -> name:Base.string -> Indexing.static_symbol Base.list -> t -> optimized

Printing

val code_hum_margin : Base.int Base.ref

val fprint_function_header : ?name:Base.string -> ?static_indices:Indexing.static_symbol Base.list -> Base.unit -> Stdlib.Format.formatter -> Base.unit

val get_ident_within_code : ?no_dots:Base.bool -> t Base.array -> Tnode.t -> Base.string

val fprint_hum : ?name:Base.string -> ?static_indices:Indexing.static_symbol Base.list -> Base.unit -> Stdlib.Format.formatter -> t -> Base.unit