Module Arrayjit.Ops

Operation types shared by all backends; and precision types.

module Lazy = Utils.Lazy

*** Precision ***

type uint8_elt = Stdlib.Bigarray.int8_unsigned_elt

type float16_elt = Stdlib.Bigarray.float32_elt

type float32_elt = Stdlib.Bigarray.float32_elt

type float64_elt = Stdlib.Bigarray.float64_elt

val float16 : (Base.float, float16_elt) Stdlib.Bigarray.kind

type ('ocaml, 'impl) precision =

| Byte : (Base.char, uint8_elt) precision

| Half : (Base.float, float16_elt) precision

| Single : (Base.float, float32_elt) precision

| Double : (Base.float, float64_elt) precision

val sexp_of_precision : 'ocaml 'impl. ('ocaml -> Sexplib0.Sexp.t) -> ('impl -> Sexplib0.Sexp.t) -> ('ocaml, 'impl) precision -> Sexplib0.Sexp.t

type prec =

| Void_prec

| Byte_prec of (Base.char, uint8_elt) precision

| Half_prec of (Base.float, float16_elt) precision

| Single_prec of (Base.float, float32_elt) precision

| Double_prec of (Base.float, float64_elt) precision

val byte : prec

val half : prec

val single : prec

val double : prec

val sexp_of_prec : prec -> Base.Sexp.t

val prec_of_sexp : Base.Sexp.t -> prec

val precision_to_string : ('ocaml, 'elt_t) precision -> string

val prec_string : prec -> string

val equal_prec : prec -> prec -> bool

val prec_in_bytes : prec -> int

val promote_prec : prec -> prec -> prec

val pack_prec : ('ocaml, 'elt_t) precision -> prec

type 'r map_prec = {

f : 'ocaml 'elt_t. ('ocaml, 'elt_t) precision -> 'r;

}

val map_prec : ?default:'a -> 'b map_prec -> prec -> 'c

val gcc_typ_of_prec : prec -> Gccjit.Type.type_kind

val cuda_typ_of_prec : prec -> string

*** Operations ***

type init_op =

| Constant_fill of {

values : Base.float Base.array;

strict : Base.bool;

}
(*
Fills in the numbers where the rightmost axis is contiguous. If strict=true, loops over the provided values.
*)

| Range_over_offsets
(*
Fills in the offset number of each cell (i.e. how many cells away it is from the beginning).
*)

| Standard_uniform
(*
Draws the values from U(0,1).
*)

| File_mapped of Base.string * prec
(*
Reads the data using Unix.openfile and Unix.map_file.
*)

Initializes or resets a array by filling in the corresponding numbers, at the appropriate precision.

val equal_init_op : init_op -> init_op -> Base.bool

val init_op_of_sexp : Sexplib0.Sexp.t -> init_op

val sexp_of_init_op : init_op -> Sexplib0.Sexp.t

type binop =

| Add

| Sub

| Mul

| Div

| ToPowOf

| Relu_gate

| Arg2

| Arg1

val binop_of_sexp : Sexplib0.Sexp.t -> binop

val sexp_of_binop : binop -> Sexplib0.Sexp.t

val compare_binop : binop -> binop -> Base.int

val equal_binop : binop -> binop -> Base.bool

type unop =

| Identity

| Relu

val unop_of_sexp : Sexplib0.Sexp.t -> unop

val sexp_of_unop : unop -> Sexplib0.Sexp.t

val compare_unop : unop -> unop -> Base.int

val equal_unop : unop -> unop -> Base.bool

val neutral_elem : binop -> float

Either the left-neutral or right-neutral element of the operation. Unspecified if the operation does not have a neutral element.

val interpret_binop : binop -> Base.Float.t -> Base.Float.t -> Base.Float.t

val interpret_unop : unop -> Base.Float.t -> Base.Float.t

val binop_C_syntax : prec -> binop -> string * string * string

val binop_cd_syntax : binop -> string

val assign_op_C_syntax : binop -> string

val assign_op_cd_syntax : initialize_neutral:bool -> binop -> string

val unop_cd_syntax : unop -> string

*** Global references ***

type voidptr = Base.unit Ctypes.ptr

val sexp_of_voidptr : unit Ctypes_static.ptr -> Base.Sexp.t

val compare_voidptr : 'a Ctypes.ptr -> 'a Ctypes.ptr -> int

val equal_voidptr : voidptr -> voidptr -> Base.bool

val ptr_to_string : 'a Ctypes.ptr -> prec -> string

type global_identifier =

| C_function of Base.string
(*
Calls a no-argument or indices-arguments C function.
*)

| External_unsafe of {

ptr : voidptr;

prec : prec;

dims : Base.int Base.array Lazy.t;

}

| Merge_buffer_unsafe
(*
Each device has at most one merge buffer, which is re-used, and re-allocated as needed, by merge operations. Using the merge buffer outside of implementing merge tasks is inherently unsafe.
*)

val sexp_of_global_identifier : global_identifier -> Sexplib0.Sexp.t

val equal_global_identifier : global_identifier -> global_identifier -> Base.bool

val compare_global_identifier : global_identifier -> global_identifier -> Base.int