Module Arrayjit.Schedulers

module Tn = Tnode

val _get_local_debug_runtime : unit -> (module Minidebug_runtime.Debug_runtime)

module Multicore (Backend : Backend_impl.For_add_scheduler) : Backend_impl.With_scheduler with type buffer_ptr = Backend.buffer_ptr

val sync_suggested_num_streams : int Base.ref

For debugging, allow Sync_scheduler(...).suggested_num_streams calls to return >1 numbers.

module Sync (Backend : Backend_impl.For_add_scheduler) : sig ... end

A minimalisitc wrapper creating backends where all calls run synchronously on the main thread. There is only one device, but an arbitrary number of streams.