Module Arrayjit.Utils

module Set_O : sig ... end

val no_ints : (Base.Int.t, Base.Int.comparator_witness) Base.Set.t

val one_int : Base.Int.t -> (Base.Int.t, Base.Int.comparator_witness) Base.Set.t

val map_merge : ('a, 'b, 'c) Base.Map.t -> ('a, 'd, 'c) Base.Map.t -> f:('e -> 'f -> 'g) -> ('a, 'h, 'c) Base.Map.t

val mref_add : ('a, 'b, 'c) Base.Map.t Base.ref -> key:'d -> data:'e -> or_:('f -> Base.unit) -> Base.unit

val mref_add_missing : ('a, 'b, 'c) Base.Map.t Base.ref -> 'd -> f:(unit -> 'e) -> Base.unit

type settings = {

mutable debug_log_from_routines : Base.bool;

mutable debug_memory_locations : Base.bool;

mutable output_debug_files_in_build_directory : Base.bool;
(*
Writes compilation related files in the build_files subdirectory of the run directory (additional files, or files that would otherwise be in temp directory). When both output_debug_files_in_build_directory = true and with_debug_level > 1, compilation should also preserve debug and line information for runtime debugging.
*)

mutable with_debug_level : Base.int;

mutable fixed_state_for_init : Base.int Base.option;

mutable print_decimals_precision : Base.int;
(*
When rendering arrays etc., outputs this many decimal digits.
*)

}

val settings_of_sexp : Sexplib0.Sexp.t -> settings

val sexp_of_settings : settings -> Sexplib0.Sexp.t

val settings : settings

val with_runtime_debug : unit -> Base.bool

val enable_runtime_debug : unit -> unit

val accessed_global_args : Base.String.t Base.Hash_set.t

val read_cmdline_or_env_var : Base.String.t -> Base.String.t option

val config_file_args : (Base.String.t, Base.String.t) Base.Hashtbl.t

val get_global_arg : default:Base.String.t -> arg_name:Base.String.t -> Base.String.t

Retrieves arg_name argument from the command line or from an environment variable, returns default if none found.

val build_file : Base.string -> Base.string

val diagn_log_file : Base.string -> Base.string

val get_debug : Base.String.t -> (module Minidebug_runtime.Debug_runtime)

module Debug_runtime : Minidebug_runtime.Debug_runtime

val union_find : equal:('a -> 'b -> bool) -> ('c, 'd, 'e) Base.Map.t -> key:'f -> rank:Base__Int.t -> 'f * Base__Int.t

val union_add : equal:('a -> 'a -> bool) -> ('b, 'b, 'c) Base.Map.t -> 'd -> 'e -> ('b, 'b, 'c) Base.Map.t

val unique_keep_first : equal:('a -> 'a -> bool) -> 'b list -> 'c Base.List.t

Filters the list keeping the first occurrence of each element.

val sorted_diff : compare:('a -> 'b -> int) -> 'c Base.List.t -> 'd list -> 'c Base.List.t

val parallel_merge : (from:Base__Int.t -> to_:Base__Int.t -> unit) -> Base.int -> Base.unit

parallel_merge merge num_devices progressively invokes the pairwise merge callback, converging on the 0th position, with from ranging from 1 to num_devices - 1, and to_ < from.

type atomic_bool = Base.bool Stdlib.Atomic.t

val sexp_of_atomic_bool : Base.bool Stdlib.Atomic.t -> Sexplib0.Sexp.t

val (!@) : 'a Stdlib.Atomic.t -> 'a

val sexp_append : elem:Base.Sexp.t -> Base.Sexp.t -> Base.Sexp.t

val sexp_mem : elem:Base.Sexp.t -> Base.Sexp.t -> bool

val sexp_deep_mem : elem:Base.Sexp.t -> Base.Sexp.t -> Base.bool

val split_with_seps : Re.re -> string -> string Base.List.t

module Lazy : sig ... end

type requirement =

| Skip

| Required

| Optional of {

callback_if_missing : Base.unit -> Base.unit;

}

val compare_requirement : requirement -> requirement -> Base.int

val requirement_of_sexp : Sexplib0.Sexp.t -> requirement

val sexp_of_requirement : requirement -> Sexplib0.Sexp.t

val get_debug_formatter : fname:Base.string -> Stdlib.Format.formatter option

exception User_error of Base.string

val header_sep : Re.re

val log_trace_tree : (module Minidebug_runtime.Debug_runtime) -> Base.String.t Base.List.t -> unit

type 'a mutable_list =

| Empty

| Cons of {

hd : 'a;

mutable tl : 'a mutable_list;

}

val equal_mutable_list : 'a. ('a -> 'a -> Base.bool) -> 'a mutable_list -> 'a mutable_list -> Base.bool

val mutable_list_of_sexp : 'a. (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a mutable_list

val sexp_of_mutable_list : 'a. ('a -> Sexplib0.Sexp.t) -> 'a mutable_list -> Sexplib0.Sexp.t

val empty : 'a mutable_list

val cons : hd:'a -> tl:'b mutable_list -> 'c mutable_list

val is_empty : 'a mutable_list -> bool

val is_cons : 'a mutable_list -> bool

val empty_val : 'a mutable_list -> unit Stdlib.Option.t

val cons_val : 'a mutable_list -> ([> `hd of 'b ] * [> `tl of 'c mutable_list ]) Stdlib.Option.t

module Variants_of_mutable_list : sig ... end

val insert : next:'a -> 'b mutable_list -> 'c mutable_list

val tl_exn : 'a mutable_list -> 'b mutable_list

type pp_file = {

f_name : Base.string;

ppf : Stdlib.Format.formatter;

finalize : Base.unit -> Base.unit;

}

val pp_file : base_name:Base.String.t -> extension:Base.String.t -> pp_file

val captured_log_prefix : string Base.ref

type captured_log_processor = {

log_processor_prefix : Base.string;

process_logs : Base.string Base.list -> Base.unit;

}

val captured_log_processors : captured_log_processor Base.list Base.ref

val add_log_processor : prefix:Base.string -> (Base.string Base.list -> Base.unit) -> Base.unit

val capture_stdout_logs : ?never_skip:Base.bool -> (unit -> 'a) -> 'b

type waiter = {

await : keep_waiting:(Base.unit -> Base.bool) -> Base.unit -> Base.bool;
(*
Returns true if the waiter was not already waiting (in another thread) and waiting was needed (keep_waiting always returned true).
*)

release_if_waiting : Base.unit -> Base.bool;
(*
Returns true if the waiter both was waiting and was not already released.
*)

is_waiting : Base.unit -> Base.bool;

finalize : Base.unit -> Base.unit;

}

Note: this waiter is meant for sequential waiting.

val waiter : name:Base.String.t -> unit -> waiter