Module Schedulers.Sync

A minimalisitc wrapper creating backends where all calls run synchronously on the main thread. There is only one device, but an arbitrary number of streams.

Parameters

module Backend : Ir.Backend_impl.For_add_scheduler

Signature

include module type of struct include Backend end

val config : Ir.Backend_intf.config

include Ir.Backend_impl.No_device_buffer_and_copying

include Ir.Backend_intf.Alloc_buffer with type stream := Base.unit

include Ir.Backend_intf.Buffer

include sig ... end

val use_host_memory : (size_in_bytes:Base.int -> Base.unit Ctypes.ptr -> Backend.buffer_ptr) Base.option

val buffer_to_buffer : dst:Backend.buffer_ptr -> src:Backend.buffer_ptr -> size_in_bytes:Base.int -> Base.unit

val host_to_buffer : Ir.Ndarray.t -> dst:Backend.buffer_ptr -> Base.unit

val buffer_to_host : Ir.Ndarray.t -> src:Backend.buffer_ptr -> Base.unit

module Device_config : sig ... end

module Device_types : sig ... end

include sig ... end

type buffer_ptr = Device_types.buffer_ptr

type dev = Device_types.dev

val sexp_of_dev : dev -> Sexplib0.Sexp.t

type runner = Device_types.runner

val sexp_of_runner : runner -> Sexplib0.Sexp.t

type event = Device_types.event

val sexp_of_event : event -> Sexplib0.Sexp.t

val name : Base.string

type optimize_ctx = Device_types.optimize_ctx

val sexp_of_optimize_ctx : optimize_ctx -> Sexplib0.Sexp.t

val empty_optimize_ctx : Base.unit -> optimize_ctx

type nonrec device

val sexp_of_device : device -> Sexplib0.Sexp.t

type nonrec stream

val sexp_of_stream : stream -> Sexplib0.Sexp.t

type nonrec context

val sexp_of_context : context -> Sexplib0.Sexp.t

val sexp_of_buffer_ptr : Device_types.buffer_ptr -> Sexplib0.Sexp.t

type nonrec buffer

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type nonrec ctx_arrays = Device_types.buffer_ptr Base.Map.M(Ir.Tnode).t

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

val alloc_array : Ir.Ops.prec -> dims:Base.int Base.array -> Device_types.stream -> Device_types.buffer_ptr

val alloc_zeros : Ir.Ops.prec -> dims:Base.int Base.array -> Device_types.stream -> Device_types.buffer_ptr

val free_buffer : (Device_types.stream -> Device_types.buffer_ptr -> Base.unit) Base.option

val make_device : 'a -> ordinal:Base.int -> ('b, 'a, 'c, 'd) Ir.Backend_intf.device

val make_stream : ('a, 'b, 'c, 'd) Ir.Backend_intf.device -> 'c -> ('a, 'b, 'c, 'd) Ir.Backend_intf.stream_ref

val get_name : ('a, 'b, 'c, 'd) Ir.Backend_intf.stream -> string

val make_context : ?ctx_arrays:(Ir.Tnode.t, 'a, Ir.Tnode.comparator_witness) Base.Map.t -> ?optimize_ctx:optimize_ctx -> 'b -> ('a, 'b, optimize_ctx) Ir.Backend_intf.context

val make_child : ?ctx_arrays:'a Ir.Backend_intf.ctx_arrays -> ?optimize_ctx:'b -> ('a, 'c, 'b) Ir.Backend_intf.context -> ('a, 'c, 'b) Ir.Backend_intf.context

val sync : unit -> unit

val is_done : unit -> bool

val will_wait_for : 'a -> unit -> unit

val alloc_buffer : ?old_buffer:Backend.buffer -> size_in_bytes:Base.int -> 'a -> Backend.buffer

val device : device

val get_device : ordinal:Base.Int.t -> device

val num_devices : unit -> int

val suggested_num_streams : 'a -> int

val get_used_memory : 'a -> Base.int

val new_stream : ('a, 'b, unit, 'c) Ir.Backend_intf.device -> ('a, 'b, unit, 'c) Ir.Backend_intf.stream_ref

val all_work : 'a -> unit

val is_idle : 'a -> bool

val await : 'a -> unit

val static_properties : Base.Sexp.t

val schedule_task : 'a -> Ir.Task.t -> Base.unit

val get_global_debug_info : unit -> Sexplib0.Sexp.t

val get_debug_info : stream -> Sexplib0.Sexp.t