Module Ir.Assignments

The code for operating on n-dimensional arrays.

module Lazy = Utils.Lazy

module Tn = Tnode

val _get_local_debug_runtime : unit -> (module Minidebug_runtime.Debug_runtime)

type buffer =

| Node of Tn.t

| Merge_buffer of Tn.t

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type fetch_op =

| Constant of Base.float

| Imported of Ops.global_identifier

| Slice of {

batch_idx : Indexing.static_symbol;

sliced : Tn.t;

}

| Embed_symbol of Indexing.static_symbol

Resets a array by performing the specified computation or data fetching.

and t =

| Noop

| Seq of t * t

| Block_comment of Base.string * t
(*
Same as the given code, with a comment.
*)

| Accum_ternop of {

initialize_neutral : Base.bool;

accum : Ops.binop;

op : Ops.ternop;

lhs : Tn.t;

rhs1 : buffer;

rhs2 : buffer;

rhs3 : buffer;

projections : Indexing.projections Lazy.t;

}

| Accum_binop of {

initialize_neutral : Base.bool;

accum : Ops.binop;

op : Ops.binop;

lhs : Tn.t;

rhs1 : buffer;

rhs2 : buffer;

projections : Indexing.projections Lazy.t;

}

| Accum_unop of {

initialize_neutral : Base.bool;

accum : Ops.binop;

op : Ops.unop;

lhs : Tn.t;

rhs : buffer;

projections : Indexing.projections Lazy.t;

}

| Fetch of {

array : Tn.t;

fetch_op : fetch_op;

dims : Base.int Base.array Lazy.t;

}

val sexp_of_fetch_op : fetch_op -> Sexplib0.Sexp.t

val sexp_of_t : t -> Sexplib0.Sexp.t

type comp = {

asgns : t;

embedded_nodes : Base.Set.M(Tn).t;
(*
The nodes in asgns that are not in embedded_nodes need to already be in contexts linked with the comp.
*)

}

Computations based on assignments. Note: the arrayjit library makes use of, but does not produce nor verify the embedded_nodes associated to some given asgns.

val sexp_of_comp : comp -> Sexplib0.Sexp.t

val to_comp : t -> comp

val empty_comp : comp

val get_name_exn : t -> Base.String.t

val is_total : initialize_neutral:Base.bool -> projections:Indexing.projections -> Base.bool

val context_nodes : use_host_memory:'a Base.option -> t -> Tn.t_set

Returns materialized nodes in the sense of Tnode.is_in_context_force. NOTE: it must be called after compilation; otherwise, it will disrupt memory mode inference.

val sequential : t Base.List.t -> t

val sequence : comp Base.List.t -> comp

val to_low_level : t -> Low_level.t

val flatten : t -> t Base__List.t

val is_noop : t -> bool

val get_ident_within_code : ?no_dots:bool -> t -> Tn.t -> Base.String.t

val doc_hum : ?name:Base.String.t -> ?static_indices:Ir__Indexing.static_symbol Base.List.t -> unit -> t -> PPrint.document

val lower : unoptim_ll_source:(PPrint.document -> Base.unit) Base.option -> ll_source:(PPrint.document -> Base.unit) Base.option -> cd_source:(PPrint.document -> unit) option -> name:Base.String.t -> Indexing.static_symbol Base.List.t -> t -> Low_level.optimized