Module Backend_impl.Device_types

Parameters

module Device_config : Backend_intf.Device_config

Signature

include module type of struct include Device_config end

include Backend_intf.Buffer

type buffer_ptr = Device_config.buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

include sig ... end

type nonrec buffer = buffer_ptr Backend_intf.buffer

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type nonrec ctx_arrays = buffer_ptr Backend_intf.ctx_arrays

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type dev = Device_config.dev

Interface to a device driver.

val sexp_of_dev : dev -> Sexplib0.Sexp.t

type runner = Device_config.runner

Interface to a stream driver.

val sexp_of_runner : runner -> Sexplib0.Sexp.t

type event = Device_config.event

An event tracks if a stream finished computing past a particular point in its schedue. These values are used internally for scheduling across streams of the backend, and can be used for explicit scheduling.

val sexp_of_event : event -> Sexplib0.Sexp.t

val name : Base.string

type nonrec device = (buffer_ptr, dev, runner, event) Backend_intf.device

val sexp_of_device : device -> Sexplib0.Sexp.t

type nonrec stream = (buffer_ptr, dev, runner, event) Backend_intf.stream

val sexp_of_stream : stream -> Sexplib0.Sexp.t

type nonrec context = (buffer_ptr, stream) Backend_intf.context

val sexp_of_context : context -> Sexplib0.Sexp.t