Parameter Device_types_ll.Device_config

type dev

Interface to a device driver.

val sexp_of_dev : dev -> Sexplib0.Sexp.t

type runner

Interface to a stream driver.

val sexp_of_runner : runner -> Sexplib0.Sexp.t

type event

An event tracks if a device's runner finished computing past a particular point in its schedule. These values are used internally for scheduling across devices/queues of the backend, and can be used for explicit scheduling.

val sexp_of_event : event -> Sexplib0.Sexp.t

val name : Base.string