Module type Backend_impl.Backend_impl_common

Parts shared by backend implementations.

include Backend_intf.Buffer

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

include sig ... end

type nonrec buffer = buffer_ptr Backend_intf.buffer

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type nonrec ctx_arrays = buffer_ptr Backend_intf.ctx_arrays

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

val use_host_memory : (size_in_bytes:Base.int -> Base.unit Ctypes.ptr -> buffer_ptr) Base.option

If not None, the backend will read from and write to the host memory directly whenever reasonable. size_in_bytes is the size of the memory allocated on the host.

use_host_memory can only be Some on unified memory devices, like CPU and Apple Metal.