Module type Backend_types.Backend_device_common

Parts shared by both assignments-level and lowered-level backend interfaces providing streams and devices.

include Device

include Device_types

include Device_config

include Buffer

type buffer_ptr

include sig ... end

type device

val sexp_of_device : device -> Sexplib0.Sexp.t

type stream_state

val sexp_of_stream_state : stream_state -> Sexplib0.Sexp.t

type runner

val sexp_of_runner : runner -> Sexplib0.Sexp.t

type event

An event tracks if a stream finished computing past a particular point in its schedue. These values are used internally for scheduling across streams of the backend, and can be used for explicit scheduling.

val sexp_of_event : event -> Sexplib0.Sexp.t

type nonrec stream = (buffer_ptr, device, stream_state, runner, event) stream

val sexp_of_stream : stream -> Sexplib0.Sexp.t

include Alloc_buffer with type buffer_ptr := buffer_ptr and type stream := stream

include Buffer with type buffer_ptr := buffer_ptr

include sig ... end

val alloc_buffer : ?old_buffer:{buffer}540 -> size_in_bytes:Base.int -> stream -> {buffer}540

val alloc_zero_init_array : Ops.prec -> dims:Base.int Base.array -> stream -> buffer_ptr

val make_stream : device:device -> state:stream_state -> unique_name:Base.string -> runner:runner -> stream

include Backend_any_common with type buffer_ptr := buffer_ptr and type init_info := stream and type stream := stream

include Buffer with type buffer_ptr := buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val c_ptr_to_string : (buffer_ptr -> Ops.prec -> Base.string) Base.option

include sig ... end

type nonrec buffer = buffer_ptr buffer

val sexp_of_buffer : buffer -> Sexplib0.Sexp.t

type ctx_arrays = buffer_ptr Base.Map.M(Arrayjit.Tnode).t

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type context

val sexp_of_context : context -> Sexplib0.Sexp.t

val name : Base.string

val initialize : config -> Base.unit

Initializes a backend before first use. Typically does nothing if the backend is already initialized, but some backends can do some safe cleanups.

val is_initialized : Base.unit -> Base.bool

Returns false if there was no previous initialize call. If it returns false, one must call initialize before using the backend.

val init : stream -> context

val finalize : context -> Base.unit

Finalizes (just) the context.

val sync : event -> Base.unit

Blocks till the event completes, if it's not done already.

val is_done : event -> Base.bool

Whether the event completed.

val will_wait_for : context -> event -> Base.unit

Schedules waiting for the given event on the context's stream.

NOTE: it should rarely be needed to call will_wait_for explicitly, because it is typically called internally when necessary. But there is one exception, see device_to_device when into_merge_buffer=Streaming.

val get_used_memory : device -> Base.int

Returns (an upper bound of) the memory used for arrays, in bytes.

val await : stream -> Base.unit

Blocks till the stream becomes idle, i.e. synchronizes the stream.

val all_work : stream -> event

Returns the event indicating if any currently running or scheduled computations on the stream have completed.

val is_idle : stream -> Base.bool

Whether the stream is currently waiting for work.

val get_device : ordinal:Base.int -> device

val num_devices : Base.unit -> Base.int

val suggested_num_streams : device -> Base.int

The optimal number of streams for the given device to follow the config strategy passed to No_device_backend.initialize.

val new_stream : device -> stream

val get_ctx_stream : context -> stream

val get_stream_device : stream -> device

val to_ordinal : device -> Base.int

val get_name : stream -> Base.string