Module Arrayjit.Backends

module Debug_runtime = Utils.Debug_runtime

val _get_local_debug_runtime : unit -> (module Minidebug_runtime.Debug_runtime)

module type No_device_backend = sig ... end

module type Backend = sig ... end

module Multicore_backend (Backend : No_device_backend) : Backend

module Pipes_multicore_backend (Backend : No_device_backend) : Backend

val sync_suggested_num_virtual_devices : int Base.ref

For debugging, allow Sync_backend(...).suggested_num_virtual_devices calls to return >1 numbers.

module Sync_backend (Backend : No_device_backend) : sig ... end

A minimalisitc wrapper creating backends where all calls run synchronously on the main thread. There is only one physical device, but an arbitrary number of virtual devices.

val lower_assignments : ?name:Base.String.t -> 'a Indexing.bindings -> Assignments.t -> Base.String.t * Low_level.optimized

val lower_batch_assignments : ?names:Base.String.t Base.Array.t -> ?occupancy:(name:Base.String.t -> src_n:int -> bool) -> 'a Indexing.bindings -> Assignments.t Base.Array.t -> Base.String.t option Base.Array.t * Low_level.optimized option Base.Array.t

module type Simple_backend = sig ... end

val verify_prior_context : ctx_arrays:('a -> (Tnode.t Base.Hashtbl.key, 'b, 'c) Base.Map.t) -> is_in_context:('d -> Base.bool) -> prior_context:'e -> from_prior_context:Tnode.t Base.Hashtbl.key Base.List.t -> (Tnode.t, 'f) Base.Hashtbl.t Base.Array.t -> unit

module Simple_no_device_backend (Backend : Simple_backend) : No_device_backend

module C_device : No_device_backend

module Cc_backend : sig ... end

module Sync_cc_backend : sig ... end

module Pipes_cc_backend : sig ... end

module Gccjit_device : No_device_backend

module Gccjit_backend : sig ... end

module Sync_gccjit_backend : sig ... end

module Pipes_gccjit_backend : sig ... end

module Cuda_backend : Backend

val reinitialize : (module Backend) -> Backend_utils.Types.config -> Base.unit

val fresh_backend : ?backend_name:Base.String.t -> ?config:Backend_utils.Types.config -> unit -> (module Backend)

Reinitializes and returns a backend corresponding to backend_name, or if omitted, selected via the global backend setting.