Module Arrayjit.Backends

module Debug_runtime = Utils.Debug_runtime

val _get_local_debug_runtime : unit -> (module Minidebug_runtime.Debug_runtime)

val check_merge_buffer : scheduled_node:Tnode.t option -> code_node:Tnode.t option -> unit

module Multicore_backend (Backend : Backend_types.No_device_backend) : Backend_types.Backend

val sync_suggested_num_streams : int Base.ref

For debugging, allow Sync_backend(...).suggested_num_streams calls to return >1 numbers.

module Sync_backend (Backend : Backend_types.No_device_backend) : Backend_types.Backend

A minimalisitc wrapper creating backends where all calls run synchronously on the main thread. There is only one device, but an arbitrary number of streams.

val lower_assignments : ?name:Base.String.t -> 'a Indexing.bindings -> Assignments.t -> Base.String.t * Low_level.optimized

val lower_batch_assignments : ?names:Base.String.t Base.Array.t -> ?occupancy:(name:Base.String.t -> src_n:int -> bool) -> 'a Indexing.bindings -> Assignments.t Base.Array.t -> Base.String.t option Base.Array.t * Low_level.optimized option Base.Array.t

val verify_prior_context : get_array:('a -> Tnode.t Base.Hashtbl.key -> 'b Base.Option.t) -> ctx_arrays:('c -> 'a) -> is_in_context:('d -> Base.bool) -> prior_context:'c -> from_prior_context:(Tnode.t Base.Hashtbl.key, 'e) Base.Set.t -> (Tnode.t, 'd) Base.Hashtbl.t Base.Array.t -> unit

val from_prior_context_batch : Assignments.comp Base.Option.t Base.Array.t -> (Tnode.t, Tnode.comparator_witness) Base.Set.t

module Lowered_no_device_backend (Backend : Backend_types.Lowered_no_device_backend) : Backend_types.No_device_backend

module C_device : Backend_types.No_device_backend

module Cc_backend : sig ... end

module Sync_cc_backend : sig ... end

module Gccjit_device : Backend_types.No_device_backend

module Gccjit_backend : sig ... end

module Sync_gccjit_backend : sig ... end

module Lowered_backend (Device : Backend_types.Lowered_backend) : Backend_types.Backend

module Cuda_backend : Backend_types.Backend

val reinitialize : (module Backend_types.Backend) -> Arrayjit__Backend_types.Types.config -> Base.unit

Initializes the backend, and if it was already initialized, performs garbage collection.

val fresh_backend : ?backend_name:Base.String.t -> ?config:Backend_types.Types.config -> unit -> (module Backend_types.Backend)

Reinitializes and returns a backend corresponding to backend_name, or if omitted, selected via the global backend setting. See reinitialize.