Module Backends.Multicore_backend

Parameters

module Backend : Backend_types.No_device_backend

Signature

type stream

val sexp_of_stream : stream -> Sexplib0.Sexp.t

include Backend_types.Backend_common with type init_info := stream and type stream := stream

type code

val sexp_of_code : code -> Sexplib0.Sexp.t

type code_batch

val sexp_of_code_batch : code_batch -> Sexplib0.Sexp.t

type buffer_ptr

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

type context

val sexp_of_context : context -> Sexplib0.Sexp.t

type routine = context Backend_types.Types.routine

val sexp_of_routine : routine -> Sexplib0.Sexp.t

val name : Base.string

val initialize : Backend_types.Types.config -> Base.unit

Initializes a backend before first use. Typically does nothing if the backend is already initialized, but some backends can do some safe cleanups.

val is_initialized : Base.unit -> Base.bool

Returns false if there was no previous initialize call. If it returns false, one must call initialize before using the backend.

val init : stream -> context

val finalize : context -> Base.unit

Finalizes (just) the context.

val alloc_buffer : ?old_buffer:(buffer_ptr * Base.int) -> size_in_bytes:Base.int -> stream -> buffer_ptr

val compile : ?shared:Base.bool -> ?name:Base.string -> Indexing.unit_bindings -> Assignments.comp -> code

If ~shared:true (default false), the backend should prefer to do more compile work in a device-and-stream-agnostic way. If ~shared:false, the backend can opt to postpone compiling altogether until link is called, to benefit from more optimizations.

val compile_batch : ?shared:Base.bool -> ?names:Base.string Base.array -> ?occupancy:(name:Base.string -> src_n:Base.int -> Base.bool) -> Indexing.unit_bindings -> Assignments.comp Base.array -> code_batch

Unlike the ~shared parameter, compile_batch vs. compile is mostly about improving the compile time and debugging convenience by generating fewer files -- ideally does not affect execution, but there can be backend-specific differences. Only array entries for which occupancy returns true are included.

val link : context -> code -> routine

Returns the routine for the code's procedure, in a new context derived from the given context.

val link_batch : context -> code_batch -> context * routine Base.option Base.array

Returns the routines for the procedures included in the code batch. The returned context is downstream of all the returned routines.

type event

An event tracks if a stream finished computing past a particular point in its schedue. These values are used internally for scheduling across streams of the backend, and can be used for explicit scheduling.

val sync : event -> Base.unit

Blocks till the event completes, if it's not done already.

val is_done : event -> Base.bool

Whether the event completed.

val work_for : context -> Tnode.t -> event Base.option

If the tensor node is in the context, returns the event indicating if currently running or scheduled computations modifying that node on the context's stream have completed.

NOTE: work_for ctx tn, if work tracking was not yet registered for tn, will register work tracking for tn and return the all_work event for ctx's stream.

val will_wait_for : context -> event -> Base.unit

Schedules waiting for the given event on the context's stream.

NOTE: it should rarely be needed to call will_wait_for explicitly, because it is typically called internally when necessary. But there is one exception, see device_to_device when into_merge_buffer=Streaming.

val from_host : context -> Tnode.t -> Base.bool

If the tensor node is both hosted and in-context, schedules a copy from host to context and returns true, otherwise returns false. NOTE: it's the caller's responsibility to synchronize the stream (via await ctx.stream or sync (work_for ctx tn)) before the host's data is overwritten.

val to_host : context -> Tnode.t -> Base.bool

If the tensor node is both hosted and in-context, schedules a copy from context to host and returns true, otherwise returns false. NOTE: it's the caller's responsibility to synchronize the stream (via await ctx.stream or sync (work_for ctx tn)) before the host's data is read.

val device_to_device : Tnode.t -> into_merge_buffer:Backend_types.Types.merge_buffer_use -> dst:context -> src:context -> Base.bool

device_to_device tn ~into_merge_buffer ~dst ~src proceeds as follows:

If the node is absent from the src context and either it is present in the dst context or into_merge_buffer is different from No: raises an error.

If the node is absent from dst and into_merge_buffer=No: returns false.

Executes will_wait_for dst (work_for src tn).

If into_merge_buffer=No: schedules a copy of the tensor node from src to dst.

If into_merge_buffer is different from No: sets on dst the merge buffer source to the given node. If into_merge_buffer=Streaming, remembers the buffer pointer of the source node to use for streaming, without blocking. If into_merge_buffer=Copy, schedules copying from src to the merge buffer of dst's stream.

NOTE: If into_merge_buffer=Streaming, after scheduling the work on dst using the merge buffer but before scheduling work on src that modifies tn, execute will_wait_for src (all_work (get_ctx_stream dst)).

type device

val get_used_memory : device -> Base.int

Returns (an upper bound of) the memory used for arrays, in bytes.

val await : stream -> Base.unit

Blocks till the stream becomes idle, i.e. synchronizes the stream.

val all_work : stream -> event

Returns the event indicating if any currently running or scheduled computations on the stream have completed.

val is_idle : stream -> Base.bool

Whether the stream is currently waiting for work.

val get_device : ordinal:Base.int -> device

val num_devices : Base.unit -> Base.int

val suggested_num_streams : device -> Base.int

The optimal number of streams for the given device to follow the Types.config strategy passed to No_device_backend.initialize.

val new_stream : device -> stream

val get_ctx_stream : context -> stream

val get_stream_device : stream -> device

val to_ordinal : device -> Base.int

val get_name : stream -> Base.string