Module Arrayjit.Cc_backend

module Lazy = Utils.Lazy

module Debug_runtime = Utils.Debug_runtime

val name : string

val optimization_level : unit -> Base.Int.t

val compiler_command : unit -> Base.String.t

Currently unused, backend behaves as if config is always `Physical_devices_only.

type mem_properties =

| Local_only
(*
The array is only needed for a local computation, is allocated on the stack.
*)

| From_context
(*
The array has a copy allocated per-cpu-device, may or may not exist on the host.
*)

| Constant_from_host
(*
The array is read directly from the host.
*)

val mem_properties_of_sexp : Sexplib0.Sexp.t -> mem_properties

val sexp_of_mem_properties : mem_properties -> Sexplib0.Sexp.t

val equal_mem_properties : mem_properties -> mem_properties -> Base.bool

val compare_mem_properties : mem_properties -> mem_properties -> Base.int

val local_only : mem_properties

val from_context : mem_properties

val constant_from_host : mem_properties

val is_local_only : mem_properties -> bool

val is_from_context : mem_properties -> bool

val is_constant_from_host : mem_properties -> bool

val local_only_val : mem_properties -> unit Stdlib.Option.t

val from_context_val : mem_properties -> unit Stdlib.Option.t

val constant_from_host_val : mem_properties -> unit Stdlib.Option.t

module Variants_of_mem_properties : sig ... end

module Tn = Tnode

type ctx_array = Ndarray.t

val sexp_of_ctx_array : ctx_array -> Sexplib0.Sexp.t

type ctx_arrays = ctx_array Base.Map.M(Tn).t

val sexp_of_ctx_arrays : ctx_arrays -> Sexplib0.Sexp.t

type context = {

label : Base.string;

arrays : ctx_arrays;

}

val sexp_of_context : context -> Sexplib0.Sexp.t

val ctx_arrays : context -> ctx_arrays

type buffer_ptr = ctx_array

Alternative approach:

type buffer_ptr = unit Ctypes_static.ptr

let sexp_of_buffer_ptr ptr = Sexp.Atom (Ops.ptr_to_string ptr Ops.Void_prec)
let buffer_ptr ctx_array = Ndarray.get_voidptr ctx_array

val sexp_of_buffer_ptr : buffer_ptr -> Sexplib0.Sexp.t

val buffer_ptr : 'a -> 'b

val alloc_buffer : ?old_buffer:('a * int) -> size_in_bytes:int -> unit -> 'b

val to_buffer : ?rt:'a -> Tn.t -> dst:Ndarray.t -> src:context -> unit

val host_to_buffer : ?rt:'a -> Ndarray.t -> dst:Ndarray.t -> unit

val buffer_to_host : ?rt:'a -> Ndarray.t -> src:Ndarray.t -> unit

val unsafe_cleanup : ?unsafe_shutdown:bool -> unit -> Base.unit

val is_initialized : unit -> bool

val initialize : unit -> Base.unit

val finalize : 'a -> unit

val init : label:Base.string -> context

type tn_info = {

tn : Tn.t;
(*
The original array.
*)

mutable ptr : Base.string;
(*
Pointer to the first value of the associated array.

if mem = Constant_from_host, the pointer to the first element of the hosted Ndarray,

if mem = From_context, either a pointer to Ndarray from context.arrays when ~shared:false, or the function parameter when ~shared:true,

if mem = Local_only, the address of the on-the-stack array.

*)

mem : mem_properties;

dims : Base.int Base.array;

size_in_elems : Base.int;

size_in_bytes : Base.int;

prec : Ops.prec;

zero_initialized : Base.bool;

}

val sexp_of_tn_info : tn_info -> Sexplib0.Sexp.t

type info_nodes = {

traced_store : Low_level.traced_store;

nodes : (Tn.t, tn_info) Base.Hashtbl.t;

used_tensors : Base.Hash_set.M(Tn).t;

get_ident : Tn.t -> Base.string;

}

val sexp_of_info_nodes : info_nodes -> Sexplib0.Sexp.t

type param_source =

| Log_file_name

| Merge_buffer

| Param_ptr of Tn.t

| Static_idx of Indexing.static_symbol

val sexp_of_param_source : param_source -> Sexplib0.Sexp.t

type procedure = {

info : info_nodes;

bindings : Indexing.unit_bindings;

name : Base.string;

result : Dl.library;

params : (Base.string * param_source) Base.list;

opt_ctx_arrays : Ndarray.t Base.Map.M(Tn).t Base.option;

expected_merge_node : Tn.t Base.option;

}

val sexp_of_procedure : procedure -> Sexplib0.Sexp.t

val expected_merge_node : procedure -> Tn.t Base.option

type ctx_nodes =

| Ctx_arrays of ctx_arrays Base.ref

| Param_ptrs of (Base.string * param_source) Base.list Base.ref

val sexp_of_ctx_nodes : ctx_nodes -> Sexplib0.Sexp.t

val pp_zero_out : Stdlib.Format.formatter -> tn_info -> unit

val get_c_ptr : Ops.prec -> Ndarray.t -> Base.String.t

val is_builtin_op : Ops.binop -> bool

val node_debug_name : tn_info -> Base.String.t

val pp_comma : Stdlib.Format.formatter -> unit -> unit

val pp_index : Stdlib.Format.formatter -> Indexing.symbol -> unit

val pp_index_axis : Stdlib.Format.formatter -> Indexing.axis_index -> unit

val pp_array_offset : Stdlib.Format.formatter -> (Indexing.axis_index Base.Array.t * int Base.Array.t) -> unit

val array_offset_to_string : (Indexing.axis_index Base.Array.t * int Base.Array.t) -> string

val prepare_node : traced_store:Low_level.traced_store -> info_nodes -> ctx_nodes -> Tn.t -> unit

val compile_main : traced_store:(Arrayjit.Low_level.Tn.t, Low_level.traced_array) Base.Hashtbl.t -> info_nodes -> Stdlib.Format.formatter -> Low_level.t -> Base.unit

val compile_globals : get_ident:(Tn.t Base.Hash_set.elt -> string) -> Stdlib.Format.formatter -> info_nodes option Base.Array.t -> unit

val compile_proc : name:string -> info_nodes -> Stdlib.Format.formatter -> Indexing.static_symbol Base.List.t -> Low_level.optimized -> (Base.String.t * param_source) Base__List.t

val prepare_nodes : info_nodes -> ctx_nodes -> Low_level.optimized -> unit

val header_sep : Re.re

val compile : name:Base.string -> opt_ctx_arrays:ctx_arrays option -> (Base.unit -> Base.unit) Indexing.bindings -> Low_level.optimized -> procedure

val compile_batch : names:Base.String.t Base.Option.t Base.Array.t -> opt_ctx_arrays: (Tn.t, ctx_array, Tn.comparator_witness) Base.Map.t Base.Option.t -> (Base.unit -> Base.unit) Indexing.bindings -> Low_level.optimized Base.option Base.array -> (Tn.t, ctx_array, Tn.comparator_witness) Base.Map.t Base.Option.t * procedure Base.Option.t Base.Array.t

val link_compiled : merge_buffer:Ndarray.t Base.option Base.ref -> context -> procedure -> context * (Indexing.static_symbol * Base.int Base.ref) Base.List.t * Tn.task * Base.string