Module Arrayjit.Utils

module Set_O : sig ... end

val no_ints : (Base.Int.t, Base.Int.comparator_witness) Base.Set.t

val one_int : Base.Int.t -> (Base.Int.t, Base.Int.comparator_witness) Base.Set.t

val map_merge : ('a, 'b, 'c) Base.Map.t -> ('a, 'd, 'c) Base.Map.t -> f:('e -> 'f -> 'g) -> ('a, 'h, 'c) Base.Map.t

val mref_add : ('a, 'b, 'c) Base.Map.t Base.ref -> key:'d -> data:'e -> or_:('f -> Base.unit) -> Base.unit

val mref_add_missing : ('a, 'b, 'c) Base.Map.t Base.ref -> 'd -> f:(unit -> 'e) -> Base.unit

type settings = {

mutable debug_log_from_routines : Base.bool;

mutable debug_memory_locations : Base.bool;

mutable output_debug_files_in_run_directory : Base.bool;

mutable with_debug : Base.bool;

mutable fixed_state_for_init : Base.int Base.option;

mutable print_decimals_precision : Base.int;
(*
When rendering arrays etc., outputs this many decimal digits.
*)

}

val settings_of_sexp : Sexplib0.Sexp.t -> settings

val sexp_of_settings : settings -> Sexplib0.Sexp.t

val settings : settings

val accessed_global_args : Base.String.t Base.Hash_set.t

val read_cmdline_or_env_var : Base.String.t -> Base.String.t option

val config_file_args : (Base.String.t, Base.String.t) Base.Hashtbl.t

val get_global_arg : default:Base.String.t -> arg_name:Base.String.t -> Base.String.t

Retrieves arg_name argument from the command line or from an environment variable, returns default if none found.

val get_debug : Base.String.t -> (module Minidebug_runtime.Debug_runtime)

module Debug_runtime : Minidebug_runtime.Debug_runtime

val union_find : equal:('a -> 'b -> bool) -> ('c, 'd, 'e) Base.Map.t -> key:'f -> rank:Base__Int.t -> 'f * Base__Int.t

val union_add : equal:('a -> 'a -> bool) -> ('b, 'b, 'c) Base.Map.t -> 'd -> 'e -> ('b, 'b, 'c) Base.Map.t

val unique_keep_first : equal:('a -> 'a -> bool) -> 'b list -> 'c Base.List.t

Filters the list keeping the first occurrence of each element.

val sorted_diff : compare:('a -> 'b -> int) -> 'c Base.List.t -> 'd list -> 'c Base.List.t

val parallel_merge : (from:Base__Int.t -> to_:Base__Int.t -> unit) -> Base.int -> Base.unit

parallel_merge merge num_devices progressively invokes the pairwise merge callback, converging on the 0th position, with from ranging from 1 to num_devices - 1, and to_ < from.

type waiter = {

await : Base.unit -> Base.unit;

release : Base.unit -> Base.unit;

finalize : Base.unit -> Base.unit;

}

val waiter : unit -> waiter

val sexp_append : elem:Base.Sexp.t -> Base.Sexp.t -> Base.Sexp.t

val sexp_mem : elem:Base.Sexp.t -> Base.Sexp.t -> bool

val sexp_deep_mem : elem:Base.Sexp.t -> Base.Sexp.t -> Base.bool

val split_with_seps : Re.re -> string -> string Base.List.t

module Lazy : sig ... end

type requirement =

| Skip

| Required

| Optional of {

callback_if_missing : Base.unit -> Base.unit;

}

val compare_requirement : requirement -> requirement -> Base.int

val requirement_of_sexp : Sexplib0.Sexp.t -> requirement

val sexp_of_requirement : requirement -> Sexplib0.Sexp.t

val get_debug_formatter : fname:Base.string -> Stdlib.Format.formatter option

exception User_error of Base.string